Add semgrep-findings call to the MCP server
This commit adds a new semgrep findings function which allows the
existing mcp server to get all findings from the API for a sepcific
deployemnt.

It's currently not possible to specify a deployment slug as it's
inferred from the `/deployments` endpoint by using the existing token.
It is currently assumed that most, if not all, users will have single
deployment per token.

This commit also adds a models.py module, which contains all the
pydantic models. New models were added for the various types
needed/returned by the findings endpoint.

Implementation is currently kept simple, use of some global variables,
and based on functions. Further abstractions are possible in the future,
once the main functionality is in.