chore: error message on using rpc when you can't (#148)
## What:
When you don't have an app token set, we will still create the process of the Semgrep daemon, but it will have erred. This stderr message gets forwarded to our logs of the server, but the MCP has no idea what happened.

THis PR makes it so we tell the MCP that the RPC server likely isn't running, so it at least knows what's up.

In the future, we should really make the daemon err out when it fails, and have the MCP middleware interpret that result.

## Test plan:
I tested both scans work when app token set, that `semgrep_scan` works when no app token, and that `semgrep_scan_rpc` returns this error message to the agent when no app token.