fix: return correct parent span in tools (#171)
`start_span`​ starts a span without setting it as the current span, while `start_as_current_span`​ starts a span AND set it as the current span. This PR fixes the way we create a new span so we get the tool span instead of the top-level span when we call `get_current_span()` in tools.

## Test plan:

You get the tool span when you call `get_current_span`​. Also in a PR stacked on top of this one, you can see attributes attaching to the correct current span.